longxia-blog/requirements.txt
# 龙虾博客 - Python 依赖

# RSS 解析
feedparser>=6.0.0

# 网页抓取
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# YAML 配置
pyyaml>=6.0

# 播客 RSS
# (使用 xml.etree.ElementTree, 内置)

# 可选: PDF 解析
# pdfplumber>=0.9.0

# 可选: Word 解析
# python-docx>=0.8.11
